+++ src/ray.c
@@ -7,7 +7,7 @@ Ray ray_create(Vec3 ori, Vec3 dir) {
     return ray;
 }
 
-Vec3 ray_point_at_param(Ray r, float t) {
+Vec3 ray_point_at_param(Ray r, double t) {
     Vec3 vec = vec3_add(r.origin, vec3_mul_esc(r.direction, t));
 
     return vec;
